feat: Add Rust workspace and a very basic Lambda function
Currently, this function logs a few pieces of data to check and
make sure it is properly receiving mail. I'm using a Rust workspace
because I plan on having at least 2 functions - one for folks to
assign tasks to themselves, and another for a task for the group
to be triaged.