feat: Use `mailparse` to pull out email data we care about
The handy mailparse library can parse the raw email that's saved
to S3, and then give us convenient handles to get the data we want.
Namely, for starters we want the sender so we know who to assign
the task, the subject for the task name, and the plain text body
for the task note/page.